Add a simple levenshtein distance calculation
This will let the user make some typos when typo a shortcode. This is nice and almost necessary since there's no tab-autocomplete.

I did some basic optimization over the gist shown to me as how to do it in lua:
https://gist.github.com/Badgerati/3261142
This is still hitchy, so I enabled it for singleplayer only.

This is less of a good implementation and more that I spent the day bashing my head against the algorithm to better understand it until I could improve it at all.